refactor(backend): reorganize structure and add prettier
Moved test files to backend/tests/ to declutter root. Added Prettier configuration and scripts.
