+++ backend/package.json
@@ -10,7 +10,8 @@
     "test": "NODE_OPTIONS=--experimental-vm-modules jest",
     "test:watch": "NODE_OPTIONS=--experimental-vm-modules jest --watch",
     "test:coverage": "NODE_OPTIONS=--experimental-vm-modules jest --coverage",
-    "lint": "eslint ."
+    "lint": "eslint .",
+    "format": "prettier --write ."
   },
   "keywords": [
     "asset-management",
@@ -42,6 +43,7 @@
     "@jest/globals": "^29.7.0",
     "eslint": "^9.39.2",
     "jest": "^29.7.0",
+    "prettier": "^3.0.0",
     "supertest": "^6.3.3"
   },
   "engines": {
