+++ backend/tests/email-templates-seeding.test.js
@@ -9,7 +9,7 @@ import { describe, it, expect, beforeAll, afterAll } from '@jest/globals';
 import { rmSync, mkdirSync } from 'fs';
 import { join } from 'path';
 import Database from 'better-sqlite3';
-import { assetDb, emailTemplateDb } from './database.js';
+import { assetDb, emailTemplateDb } from '../database.js';
 
 const TEST_DB_DIR = join(process.cwd(), 'test-data-email-templates');
 const TEST_DB_PATH = join(TEST_DB_DIR, 'assets.db');
