+++ backend/tests/footer-label-migration.test.js
