+++ backend/tests/manager-read-authorization.test.js
@@ -1,9 +1,9 @@
 import { describe, it, expect, beforeAll, afterAll } from '@jest/globals';
-import { userDb, companyDb, attestationCampaignDb } from './database.js';
-import { generateToken } from './auth.js';
+import { userDb, companyDb, attestationCampaignDb } from '../database.js';
+import { generateToken } from '../auth.js';
 import request from 'supertest';
 import express from 'express';
-import { authenticate, authorize } from './auth.js';
+import { authenticate, authorize } from '../auth.js';
 
 // Create minimal express app with the endpoints for testing
 const app = express();
